requirements.txt
streamlit
watchdog
pymupdf
python-docx
flask
flask-cors